+++ src/main/kotlin/org/consoleapp/com/models/MemberModel.kt
@@ -1,9 +1,14 @@
 package org.consoleapp.com.models
 
-data class MemberModel(var id: Long = 0,
-                       var firstName: String = "",
-                       var lastName: String = "",
-                       var dateOfBirth: String = "",
-                       var email: String = "",
-                       var phoneNumber: String = ""
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+
+data class MemberModel(
+    var id: Int = 0,
+    var firstName: String = "",
+    var lastName: String = "",
+    var dateOfBirth : String = "",
+    //var dateOfBirth: LocalDate = LocalDate.parse("2020-01-10", DateTimeFormatter.ISO_DATE),
+    var email: String = "",
+    var phoneNumber: String = ""
 )
